Expose network availability as a StateFlow and rename class to NetworkMonitor
