+++ demo-app/src/main/kotlin/io/getstream/video/android/util/NetworkMonitor.kt
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.video.android.util
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.Network
+import android.net.NetworkCapabilities
+import io.getstream.log.taggedLogger
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+
+class NetworkMonitor(context: Context) {
+    private val connectivityManager = context.getSystemService(
+        Context.CONNECTIVITY_SERVICE,
+    ) as ConnectivityManager
+    private val logger by taggedLogger("NetworkConnectivity")
+
+    private val _isNetworkAvailable = MutableStateFlow(true)
+    val isNetworkAvailable = _isNetworkAvailable.asStateFlow()
+
+    init {
+        // Check initial network state
+        _isNetworkAvailable.value = connectivityManager.getNetworkCapabilities(
+            connectivityManager.activeNetwork,
+        )?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) == true
+
+        connectivityManager.registerDefaultNetworkCallback(getNetworkCallback())
+    }
+
+    private fun getNetworkCallback() = object : ConnectivityManager.NetworkCallback() {
+        override fun onAvailable(network: Network) {
+            super.onAvailable(network)
+            _isNetworkAvailable.value = true
+            logger.i { "Network available" }
+        }
+
+        override fun onLost(network: Network) {
+            super.onLost(network)
+            _isNetworkAvailable.value = false
+            logger.i { "Network lost" }
+        }
+
+        override fun onUnavailable() {
+            super.onUnavailable()
+            _isNetworkAvailable.value = false
+            logger.i { "Network unavailable" }
+        }
+    }
+}
